feat: add monitoring package
