+++ changelog.d/2025.09.05.23.05.31.added.md
@@ -1,0 +1,1 @@
+- add monitoring package combining metrics and rate limiting utilities
