+++ docs/monitoring.md
@@ -1,0 +1,3 @@
+# Monitoring
+
+Provides a lightweight TokenBucket rate limiter and Prometheus metrics helpers.
