+++ packages/discord-rest/package.json
@@ -14,7 +14,7 @@
   },
   "dependencies": {
     "@promethean/agent": "workspace:*",
-    "@promethean/rate": "workspace:*",
+    "@promethean/monitoring": "workspace:*",
     "@promethean/security": "workspace:*",
     "@promethean/platform": "workspace:*"
   }
