--- packages/metrics/package.json
@@ -1,63 +1,0 @@
-{
-    "name": "@promethean/metrics",
-    "version": "0.0.1",
-    "private": true,
-    "type": "module",
-    "main": "dist/index.cjs",
-    "types": "dist/index.d.ts",
-    "exports": {
-        ".": {
-            "types": "./dist/index.d.ts",
-            "import": "./dist/index.js",
-            "require": "./dist/index.cjs"
-        },
-        "./dist/*": "./dist/*",
-        "./*": "./dist/*"
-    },
-    "files": [
-        "dist"
-    ],
-    "scripts": {
-        "build": "tsc",
-        "clean": "rimraf dist",
-        "typecheck": "tsc -p tsconfig.json --noEmit",
-        "test": "pnpm run build && ava --config ../../config/ava.config.mjs",
-        "lint": "pnpm exec biome lint .",
-        "lisp": "node dist/cli/lisp.js",
-        "coverage": "pnpm run build && c8 ava --config ../../config/ava.config.mjs",
-        "test:markdown": "pnpm run build && ava --config ../../config/ava.config.mjs dist/tests/markdown.*.test.js",
-        "format": "pnpm exec biome format --write ."
-    },
-    "module": "dist/index.js",
-    "dependencies": {
-        "@types/javascript-time-ago": "^2.5.0",
-        "@types/unist": "^3.0.3",
-        "body-parser": "^1.20.2",
-        "chromadb": "^3.0.14",
-        "duckduckgo-search": "^1.0.7",
-        "express": "^4.19.2",
-        "gray-matter": "^4.0.3",
-        "javascript-time-ago": "^2.5.11",
-        "mongodb": "^6.18.0",
-        "ollama": "^0.5.17",
-        "prom-client": "^15.1.0",
-        "remark-gfm": "^4.0.1",
-        "remark-parse": "^11.0.0",
-        "remark-stringify": "^11.0.0",
-        "sucrase": "^3.35.0",
-        "unified": "^11.0.5",
-        "unist": "^0.0.1",
-        "unist-util-to-string-with-nodes": "^0.1.1",
-        "unist-util-visit": "^5.0.0",
-        "uuid": "^11.1.0",
-        "ws": "^8.18.0",
-        "yaml": "^2.5.1",
-        "zod": "^3.23.8"
-    },
-    "devDependencies": {
-        "@types/body-parser": "^1.19.5",
-        "@types/estree": "^1.0.5",
-        "@types/express": "^4.17.21",
-        "@types/ws": "^8.5.9"
-    }
-}
