+++ packages/monitoring/ava.config.mjs
@@ -1,0 +1,2 @@
+import base from '../../config/ava.config.mjs';
+export default { ...base, files: ['tests/**/*.test.js'] };
