+++ packages/monitoring/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@promethean/monitoring",
+  "version": "0.0.1",
+  "private": true,
+  "type": "module",
+  "main": "dist/index.cjs",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs"
+    },
+    "./dist/*": "./dist/*",
+    "./*": "./dist/*"
+  },
+  "files": ["dist"],
+  "scripts": {
+    "build": "tsc",
+    "clean": "rimraf dist",
+    "typecheck": "tsc -p tsconfig.json --noEmit",
+    "test": "pnpm run build && ava tests/**/*.test.js",
+    "lint": "pnpm exec biome lint .",
+    "coverage": "pnpm run build && c8 ava tests/**/*.test.js",
+    "format": "pnpm exec biome format --write ."
+  },
+  "module": "dist/index.js",
+  "dependencies": {
+    "prom-client": "^15.1.0",
+    "@promethean/test-utils": "workspace:*"
+  }
+}
