+++ packages/monitoring/src/index.ts
@@ -1,1 +1,2 @@
+export { metrics } from './prom.js';
 export { TokenBucket } from './limiter.js';
