+++ packages/monitoring/tests/limiter.test.js
@@ -21,3 +21,14 @@ test('TokenBucket refills over time', async (t) => {
     await sleep(1100);
     t.true(bucket.tryConsume());
 });
+
+test('TokenBucket limits and refills with capacity 2', async (t) => {
+    const bucket = new TokenBucket({ capacity: 2, refillPerSec: 1 });
+    t.true(bucket.tryConsume());
+    t.true(bucket.tryConsume());
+    t.false(bucket.tryConsume());
+    const deficit = bucket.deficit();
+    t.true(deficit > 0);
+    await sleep(1100);
+    t.true(bucket.tryConsume());
+});
