+++ packages/monitoring/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "../../config/tsconfig.base.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "dist",
+    "composite": true,
+    "declaration": true,
+    "moduleResolution": "nodenext"
+  },
+  "include": ["src/**/*"],
+  "references": []
+}
