--- packages/rate/.eslintrc.cjs
@@ -1,1 +1,0 @@
-module.exports = { extends: ['../../../config/.eslintrc.base.cjs'] };
