--- packages/rate/project.json
@@ -1,33 +1,0 @@
-{
-    "name": "ts-rate",
-    "root": "shared/ts/rate",
-    "sourceRoot": "shared/ts/rate/src",
-    "targets": {
-        "build": {
-            "executor": "nx:run-commands",
-            "options": {
-                "command": "tsc -b"
-            },
-            "outputs": ["{projectRoot}/dist"]
-        },
-        "typecheck": {
-            "executor": "nx:run-commands",
-            "options": {
-                "command": "tsc -p tsconfig.json --noEmit"
-            }
-        },
-        "test": {
-            "executor": "nx:run-commands",
-            "options": {
-                "command": "ava"
-            }
-        },
-        "lint": {
-            "executor": "nx:run-commands",
-            "options": {
-                "command": "eslint ."
-            }
-        }
-    },
-    "tags": ["scope:shared-ts"]
-}
