+++ packages/ws/package.json
@@ -31,7 +31,7 @@
     "module": "dist/index.js",
     "dependencies": {
         "@promethean/event": "workspace:*",
-        "@promethean/rate": "workspace:*",
+        "@promethean/monitoring": "workspace:*",
         "@types/javascript-time-ago": "^2.5.0",
         "@types/unist": "^3.0.3",
         "body-parser": "^1.20.2",
