+++ packages/ws/src/shims.d.ts
@@ -1,2 +1,1 @@
-declare module '@promethean/rate/limiter.js';
 declare module '@promethean/event/types.js';
